requirements.txt
azure-storage-blob>=12.19.0
azure-ai-documentintelligence>=1.0.0b1
azure-identity>=1.15.0
python-dotenv>=1.0.0
tqdm>=4.66.0
